Exercises 4.21. - 4.23.

+++ part4/blogilista/controllers/blogs.js
@@ -1,11 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 require('express-async-errors')
 
 blogsRouter.get('/', async (request, response) => {
-  
   const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
@@ -23,12 +21,18 @@ blogsRouter.post('/', async (request, response) => {
   
   const body = request.body
   //const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  /* const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
 
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(decodedToken.id) */
+
+  const user = request.user
+  if (!user)
+  {
+    return response.status(401).end()
+  }
 
   const blog = new Blog({
     title: body.title,
@@ -58,7 +62,19 @@ blogsRouter.post('/', async (request, response) => {
 blogsRouter.delete('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
   if (!blog){
-    response.status(400).end()
+    return response.status(400).end()
+  }
+
+  /* const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id) */
+  const user = request.user
+
+  if (!user || blog.user.toString() !== user._id.toString()){
+    return response.status(401).json({ error: 'invalid user token' })
   }
 
   await Blog.findByIdAndRemove(request.params.id)
